refactor ref path
